feat: Record for registration responses
